refactor: 코드 리뷰 반영 - AcceptanceTest 클래스 패키지 위치 변경
