+++ src/test/java/kitchenpos/menu/acceptance/MenuAcceptanceTest.java
@@ -14,7 +14,7 @@ import org.junit.jupiter.api.Test;
 
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import kitchenpos.utils.AcceptanceTest;
+import kitchenpos.AcceptanceTest;
 import kitchenpos.menugroup.acceptance.MenuGroupId;
 import kitchenpos.product.acceptance.ProductId;
