+++ src/test/java/kitchenpos/tablegroup/acceptance/TableGroupAcceptanceTest.java
@@ -13,7 +13,7 @@ import org.junit.jupiter.api.Test;
 
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import kitchenpos.utils.AcceptanceTest;
+import kitchenpos.AcceptanceTest;
 import kitchenpos.order.acceptance.OrderId;
 import kitchenpos.table.acceptance.OrderTableId;
